Make a clear distinction between the app and its system
They are not the same. Logically, the app contains the system (of stateful components).
Also, there are to logically separate configs — for the app as a whole (a regular one)
and for the Integrant system (Integrant-specific config map) — though these are usually
combined/conflated into a single one. See the next commit for the app config setting.